fix: make sure the PR number in a dashboard is sorted numerically
and not as a string (which would put e.g. 1200 before 900).
To do so, do not add a hash in front of the PR number.
That breaks auto-detecting and sorting the column's data type.

A more advanced fix would be to define a custom data type "PR number",
with a custom sorting function which strips the HTML tags, the the leading #
and then sorts the remainder as numbers. See e.g. the example at
https://datatables.net/reference/api/DataTable.type()
for how to define a custom type.

Another option would be a custom renderer (having the underlying *data* be
just a number, but *displaying* it with a leading hash).
We leave these options to a future commit.