refactor: collect lists of all PRs in a central place
This will make it easier to add a 'cross-cutting' dashboard.
In doing these changes, we
- move the Label and BasicPRInformation types up (to enable type hints)
- make _extract_prs only take one single data file (as we extract from each file
individually, and merge them later)
- make print_dashboard and the statistics computation operate on the extracted
PR info, thus centralising the JSON reading operation (more).
- As a side effect, this makes the 'bad title' dashboard display diff information
when available (which is certainly nice).

The order of the "bad title" and "unlabelled" dashboards was inconsistent
between the table of contents and the actual page. We change the TOC
to match the order on the page: with the recently introduced auto-labelling
of PRs on mathlib, unlabelled PRs should become rare.