refactor: remove separate API calls for stale PRs ready to merge
Instead, just have one dashboard for all PRs not updated in a day,
and filter this dashboard instead. This should fix the duplication bug
as a side effect.
The new dashboard could also be removed, with further post-processing:
we leave that as a next step. Let us first remove as many easy API calls
as possible.

An unfortunate side effect is that this makes running tests harder,
as a file for stale PRs needs to be provided. Let us hope this problem
resolves itself soon.