chore: move printing of the on_the_queue page next to the other pages
And make its set-up consistent with the others: name write_, and hard-code
the output file as well. *If* I want to allow changing this, I should do it
for all webpages at once. (And I'm not sure if instruction the user to copy
and diff all *.html files instead is not a better approach.)