refactor: more precise and less hacky sorting of the total time in review
- Refactor the 'total time in state X' function to return this timespan
as both a relativedelta (for 'N months ago' or 'M hours ago' formatting)
and a timedelta (which has only days and seconds, so is easier to sort).
- Use the timedelta information for sorting the 'total time in state X'
information: this is easier/yields slightly shorter code, and also less hacky:
no more need to assume a month has 31 days or a year has 366 days.