chore: audit list of PRs with exactly 100 commits
Currently, all of these have incomplete data.
Some of them have many more than 100 commits, so should not be redownloaded
for this purpose either.# mark a few PRs with LOTS of commits as stubborn already

Sort and deduplicate the file of stubborn PRs. Two PRs have lots of data,
but the current metadata finally worked out fine: so we let it be, for now.