feat(UX): show added and deleted lines in colour
Currently, we use green for additions and red for deletions:
this is not ideal (e.g., because of the positive/negative connotations
or for red-green-blind users), for far-spread.

Supporting other colour schemes is possible; I'm just not sure how
to do this without imposing this choice on everybody. Contributions welcome!

Fixes #88.